django integrations

+++ resources/server/models.py
@@ -1,6 +1,8 @@
+# -*- coding: utf-8 -*-
+
 from django.db import models
 
 
 class Server(models.Model):
-    discord_id = models.IntegerField(default=0,  null=False, blank=False, db_index=True)
+    discord_id = models.BigIntegerField(default=0, null=False, blank=False, db_index=True)
     enabled = models.BooleanField(default=True)
